Update property model to use decimal field for insurance, create property create and update views and templates

+++ properties/models.py
@@ -28,7 +28,7 @@ class Property(models.Model):
     property_type = models.CharField(max_length=20, choices=PROPERTY_TYPE_CHOICES, default=COMMERCIAL, verbose_name='Property Type')
     square_feet = models.IntegerField(default=0)
     taxes = models.IntegerField(default=0)
-    insurance = models.IntegerField(default=0)
+    insurance = models.DecimalField(max_digits=12, decimal_places=2, default=0.0)
     is_deleted = models.BooleanField(default=False)
     created_on = models.DateTimeField(auto_now_add=True)
     modified_on = models.DateTimeField(auto_now=True)
